Changed the way we work with databases
Database creation is now included in the module - Models
Database migrations are brought to the module - Manage
